+++ pkg/component/gardener/dashboard/deployment.go
@@ -260,7 +260,7 @@ func (g *gardenerDashboard) deployment(
 							VolumeSource: corev1.VolumeSource{
 								Secret: &corev1.SecretVolumeSource{
 									SecretName:  secretNameServerCert,
-									DefaultMode: new(0640),
+									DefaultMode: new(int32(0640)),
 								},
 							},
 						},
